#46 Simplifying internal structure of resource-po…
…ol library, and also not exposing internal members.
