+++ resource-pool/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ty-ras-extras/resource-pool-fp-ts",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "author": {
     "name": "Stanislav Muhametsin",
     "email": "346799+stazz@users.noreply.github.com",
